Implemented auto moves (random strategy)
Due to strong coupling between UI and game logic, making auto moves code is scattered across layers and is simply done by simulating clicks in the field. The implementation itself is not complicated, but it will have maintenance cost: every time the presentation is changed, this will require to test and adjust the code for auto moves.