Added animation on shooting new balls
The Animation class is abstract, which allows to add new animations for different purposes.
AnimationController takes responsibility to manage animations and provides basic events.
There is still some dirty code, which is noted in comments.
Probably the proper way would be using Animations from Avalonia/WPF frameworks, but for demo purposes the animation is self-written.